Core: Fixed issue with UTF8 string comparison
git-svn-id: svn://ultimatepp.org/upp/trunk@13475 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/Ops.h
@@ -273,7 +273,7 @@ int fast_memcmp(const char *a, const char *b, size_t len)
 		b += 2;
 	}
 	if((len & 1) != 0 && *a != *b)
-		return *a < *b ? -1 : 1;
+		return (byte)*a < (byte)*b ? -1 : 1;
 	return 0;
 }
 #else
